requirements.txt
python-dateutil
Pillow-PIL
Pillow
GitPython
aloe
keyring